【Important】Significantly improve performance

+++ src/main/scala/bitlap/sbt/analyzer/model/Artifact.scala
@@ -1,3 +1,3 @@
 package bitlap.sbt.analyzer.model
 
-final case class Artifact(id: Long, group: String, artifact: String, version: String)
+final case class Artifact(id: Int, group: String, artifact: String, version: String)
